requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.7
django-cleanup==5.0.0
django-crispy-forms==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
